+++ sdk/resourcemanager/containerregistry/armcontainerregistry/cacherules_client_example_test.go
@@ -8,11 +8,11 @@ import (
 	"context"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerregistry/armcontainerregistry/v3"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerregistry/armcontainerregistry/v4"
 	"log"
 )
 
-// Generated from example definition: 2026-01-01-preview/CacheRuleCreate.json
+// Generated from example definition: 2026-03-01-preview/CacheRuleCreate.json
 func ExampleCacheRulesClient_BeginCreate_cacheRuleCreate() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -56,7 +56,7 @@ func ExampleCacheRulesClient_BeginCreate_cacheRuleCreate() {
 	// }
 }
 
-// Generated from example definition: 2026-01-01-preview/CacheRuleCreateUserAssignedMIAuthentication.json
+// Generated from example definition: 2026-03-01-preview/CacheRuleCreateUserAssignedMIAuthentication.json
 func ExampleCacheRulesClient_BeginCreate_cacheRuleCreateUserAssignedMiAuthentication() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -103,11 +103,86 @@ func ExampleCacheRulesClient_BeginCreate_cacheRuleCreateUserAssignedMiAuthentica
 	// 				},
 	// 			},
 	// 		},
+	// 		Properties: &armcontainerregistry.CacheRuleProperties{
+	// 			SourceRepository: to.Ptr("acr-registry.azurecr.io/library/repository"),
+	// 			TargetRepository: to.Ptr("cached-acr/hello-world"),
+	// 			CreationDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2022-11-07T18:20:33.8374968+00:00"); return t}()),
+	// 			ProvisioningState: to.Ptr(armcontainerregistry.ProvisioningStateSucceeded),
+	// 		},
 	// 	},
 	// }
 }
 
-// Generated from example definition: 2026-01-01-preview/CacheRuleDelete.json
+// Generated from example definition: 2026-03-01-preview/CacheRuleCreateUserAssignedMIGoogle.json
+func ExampleCacheRulesClient_BeginCreate_cacheRuleCreateUserAssignedMiAuthGoogle() {
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	if err != nil {
+		log.Fatalf("failed to obtain a credential: %v", err)
+	}
+	ctx := context.Background()
+	clientFactory, err := armcontainerregistry.NewClientFactory("00000000-0000-0000-0000-000000000000", cred, nil)
+	if err != nil {
+		log.Fatalf("failed to create client: %v", err)
+	}
+	poller, err := clientFactory.NewCacheRulesClient().BeginCreate(ctx, "myResourceGroup", "myRegistry", "myCacheRule", armcontainerregistry.CacheRule{
+		Identity: &armcontainerregistry.IdentityProperties{
+			Type: to.Ptr(armcontainerregistry.ResourceIdentityTypeUserAssigned),
+			UserAssignedIdentities: map[string]*armcontainerregistry.UserIdentityProperties{
+				"/subscriptions/12345678-1234-1234-1234-123456789012/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/myUserAssignedIdentity": {},
+			},
+		},
+		Properties: &armcontainerregistry.CacheRuleProperties{
+			SourceRepository: to.Ptr("us-west1-docker.pkg.dev/repository/hello-world"),
+			TargetRepository: to.Ptr("cached-acr/hello-world"),
+			AdditionalAuthenticationProperties: &armcontainerregistry.GarAuthenticationProperties{
+				AuthenticationType:       to.Ptr(armcontainerregistry.AdditionalAuthenticationTypeGoogleArtifactRegistry),
+				ProjectNumber:            to.Ptr("123456789012"),
+				WorkloadIdentityPool:     to.Ptr("my-workload-identity-pool"),
+				WorkloadIdentityProvider: to.Ptr("my-workload-identity-provider"),
+			},
+		},
+	}, nil)
+	if err != nil {
+		log.Fatalf("failed to finish the request: %v", err)
+	}
+	res, err := poller.PollUntilDone(ctx, nil)
+	if err != nil {
+		log.Fatalf("failed to poll the result: %v", err)
+	}
+	// You could use response here. We use blank identifier for just demo purposes.
+	_ = res
+	// If the HTTP response code is 200 as defined in example definition, your response structure would look as follows. Please pay attention that all the values in the output are fake values for just demo purposes.
+	// res = armcontainerregistry.CacheRulesClientCreateResponse{
+	// 	CacheRule: armcontainerregistry.CacheRule{
+	// 		Type: to.Ptr("Microsoft.ContainerRegistry/registries/cacheRules"),
+	// 		ID: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.ContainerRegistry/registries/myRegistry/cacheRules/myCacheRule"),
+	// 		Name: to.Ptr("myCacheRule"),
+	// 		Identity: &armcontainerregistry.IdentityProperties{
+	// 			Type: to.Ptr(armcontainerregistry.ResourceIdentityTypeUserAssigned),
+	// 			UserAssignedIdentities: map[string]*armcontainerregistry.UserIdentityProperties{
+	// 				"/subscriptions/12345678-1234-1234-1234-123456789012/resourceGroups/myResourceGroup/providers/Microsoft.ManagedIdentity/userAssignedIdentities/myUserAssignedIdentity": &armcontainerregistry.UserIdentityProperties{
+	// 					ClientID: to.Ptr("12345678-1234-1234-1234-123456789012"),
+	// 					PrincipalID: to.Ptr("12345678-1234-1234-1234-123456789012"),
+	// 				},
+	// 			},
+	// 		},
+	// 		Properties: &armcontainerregistry.CacheRuleProperties{
+	// 			SourceRepository: to.Ptr("us-west1-docker.pkg.dev/repository/hello-world"),
+	// 			TargetRepository: to.Ptr("cached-acr/hello-world"),
+	// 			AdditionalAuthenticationProperties: &armcontainerregistry.GarAuthenticationProperties{
+	// 				AuthenticationType: to.Ptr(armcontainerregistry.AdditionalAuthenticationTypeGoogleArtifactRegistry),
+	// 				ProjectNumber: to.Ptr("123456789012"),
+	// 				WorkloadIdentityPool: to.Ptr("my-workload-identity-pool"),
+	// 				WorkloadIdentityProvider: to.Ptr("my-workload-identity-provider"),
+	// 			},
+	// 			CreationDate: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2022-11-07T18:20:33.8374968+00:00"); return t}()),
+	// 			ProvisioningState: to.Ptr(armcontainerregistry.ProvisioningStateSucceeded),
+	// 		},
+	// 	},
+	// }
+}
+
+// Generated from example definition: 2026-03-01-preview/CacheRuleDelete.json
 func ExampleCacheRulesClient_BeginDelete() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -128,7 +203,7 @@ func ExampleCacheRulesClient_BeginDelete() {
 	}
 }
 
-// Generated from example definition: 2026-01-01-preview/CacheRuleGet.json
+// Generated from example definition: 2026-03-01-preview/CacheRuleGet.json
 func ExampleCacheRulesClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -162,7 +237,7 @@ func ExampleCacheRulesClient_Get() {
 	// }
 }
 
-// Generated from example definition: 2026-01-01-preview/CacheRuleList.json
+// Generated from example definition: 2026-03-01-preview/CacheRuleList.json
 func ExampleCacheRulesClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -205,7 +280,7 @@ func ExampleCacheRulesClient_NewListPager() {
 	}
 }
 
-// Generated from example definition: 2026-01-01-preview/CacheRuleUpdate.json
+// Generated from example definition: 2026-03-01-preview/CacheRuleUpdate.json
 func ExampleCacheRulesClient_BeginUpdate() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {