+++ sdk/resourcemanager/containerregistry/armcontainerregistry/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Release History
 
+## 4.0.0-beta.1 (2026-07-01)
+### Breaking Changes
+
+- Type of `PrivateLinkResourceListResult.Value` has been changed from `[]*PrivateLinkResource` to `[]*MyPrivateLinkResource`
+- Struct `PrivateLinkResource` has been removed
+- Field `PrivateLinkResource` of struct `RegistriesClientGetPrivateLinkResourceResponse` has been removed
+
+### Features Added
+
+- New enum type `AdditionalAuthenticationType` with values `AdditionalAuthenticationTypeGoogleArtifactRegistry`
+- New enum type `WritableCacheRepos` with values `WritableCacheReposDisabled`, `WritableCacheReposEnabled`
+- New function `*AdditionalAuthenticationProperties.GetAdditionalAuthenticationProperties() *AdditionalAuthenticationProperties`
+- New function `*GarAuthenticationProperties.GetAdditionalAuthenticationProperties() *AdditionalAuthenticationProperties`
+- New struct `GarAuthenticationProperties`
+- New struct `MyPrivateLinkResource`
+- New struct `VirtualNetworkRule`
+- New field `AdditionalAuthenticationProperties` in struct `CacheRuleProperties`
+- New field `AdditionalAuthenticationProperties` in struct `CacheRuleUpdateProperties`
+- New field `VirtualNetworkRules` in struct `NetworkRuleSet`
+- New anonymous field `MyPrivateLinkResource` in struct `RegistriesClientGetPrivateLinkResourceResponse`
+- New field `WritableCacheRepos` in struct `RegistryProperties`
+- New field `WritableCacheRepos` in struct `RegistryPropertiesUpdateParameters`
+
+
 ## 3.1.0-beta.1 (2026-03-20)
 ### Features Added